feat: Color chip 추가
