+++ core/designsystem/src/main/java/com/suwiki/core/designsystem/component/chips/SuwikiOutlinedChip.kt
@@ -1,6 +1,7 @@
-package com.suwiki.core.designsystem.component.button
+package com.suwiki.core.designsystem.component.chips
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -20,23 +21,14 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.suwiki.core.designsystem.component.button.ui.theme.F6
-import com.suwiki.core.designsystem.component.button.ui.theme.TestTheme
-import com.suwiki.core.designsystem.component.button.ui.theme.blue10
-import com.suwiki.core.designsystem.component.button.ui.theme.blue100
-import com.suwiki.core.designsystem.component.button.ui.theme.c95
-import com.suwiki.core.designsystem.component.button.ui.theme.green10
-import com.suwiki.core.designsystem.component.button.ui.theme.green100
-import com.suwiki.core.designsystem.component.button.ui.theme.orange10
-import com.suwiki.core.designsystem.component.button.ui.theme.orange100
-enum class SuwikiLabelType(val type: String) {
-  ORANGE("orange"),
-  BLUE("blue"),
-  GREEN("green"),
-}
+import com.suwiki.core.designsystem.component.chips.ui.theme.DA
+import com.suwiki.core.designsystem.component.chips.ui.theme.Main
+import com.suwiki.core.designsystem.component.chips.ui.theme.TestTheme
+import com.suwiki.core.designsystem.component.chips.ui.theme.c95
+import com.suwiki.core.designsystem.component.chips.ui.theme.white
 
 @Composable
-fun SuwikiContainedChips(type: SuwikiLabelType, text: String) {
+fun SuwikiOutlinedChips(text: String) {
   var labelState by remember { mutableStateOf(false) }
 
   val backgroundColor: Color
@@ -44,28 +36,16 @@ fun SuwikiContainedChips(type: SuwikiLabelType, text: String) {
 
   when (labelState) {
     false -> {
-      backgroundColor = F6
+      backgroundColor = DA
       contentColor = c95
     }
     true -> {
-      when (type) {
-        SuwikiLabelType.ORANGE -> {
-          backgroundColor = orange10
-          contentColor = orange100
-        }
-        SuwikiLabelType.BLUE -> {
-          backgroundColor = blue10
-          contentColor = blue100
-        }
-        SuwikiLabelType.GREEN -> {
-          backgroundColor = green10
-          contentColor = green100
-        }
-      }
+      backgroundColor = Main
+      contentColor = Main
     }
   }
 
-  SuwikiContainedChipsItem(
+  SuwikiOutlinedChipsItem(
     text = text,
     backgroundColor = backgroundColor,
     contentColor = contentColor,
@@ -74,7 +54,7 @@ fun SuwikiContainedChips(type: SuwikiLabelType, text: String) {
 }
 
 @Composable
-fun SuwikiContainedChipsItem(
+fun SuwikiOutlinedChipsItem(
   text: String,
   backgroundColor: Color,
   contentColor: Color,
@@ -83,9 +63,10 @@ fun SuwikiContainedChipsItem(
   Box(
     modifier = Modifier
       .clip(RoundedCornerShape(5.dp))
-      .background(color = backgroundColor)
       .clickable { onBtnClicked() }
-      .size(41.dp, 26.dp),
+      .size(41.dp, 26.dp)
+      .background(color = white)
+      .border(width = 1.dp, color = backgroundColor, shape = RoundedCornerShape(5.dp)),
   ) {
     Text(
       text = text,
@@ -99,13 +80,13 @@ fun SuwikiContainedChipsItem(
 
 @Preview(showBackground = true, backgroundColor = 0xFFFFFF)
 @Composable
-fun SuwikiContainedChipsPreview() {
+fun SuwikiOutlinedChipsPreview() {
   TestTheme {
     Column(
       verticalArrangement = Arrangement.Center,
       horizontalAlignment = Alignment.CenterHorizontally,
     ) {
-      SuwikiContainedChips(SuwikiLabelType.ORANGE, "label")
+      SuwikiOutlinedChips("label")
     }
   }
 }
