+++ core/designsystem/src/main/java/com/suwiki/core/designsystem/component/chips/ui/theme/TestColor.kt
@@ -1,4 +1,4 @@
-package com.suwiki.core.designsystem.component.button.ui.theme
+package com.suwiki.core.designsystem.component.chips.ui.theme
 
 import androidx.compose.ui.graphics.Color
