+++ core/designsystem/src/main/java/com/suwiki/core/designsystem/component/chips/ui/theme/TestTheme.kt
@@ -1,4 +1,4 @@
-package com.suwiki.core.designsystem.component.button.ui.theme
+package com.suwiki.core.designsystem.component.chips.ui.theme
 
 import android.app.Activity
 import android.os.Build
